move jcli actions that do not require a running node to a separate mod
