+++ src/utils/mod.rs
@@ -1,6 +1,7 @@
 pub mod blockchain;
 pub mod download;
 pub mod github;
+pub mod jcli;
 pub mod release;
 pub mod runner;
 pub mod version;
