+++ src/utils/runner.rs
@@ -3,7 +3,7 @@ use serde::{Deserialize, Serialize};
 use std::{
     io,
     net::SocketAddr,
-    path::{Path, PathBuf},
+    path::PathBuf,
     process::{Command, Stdio},
 };
 use thiserror::Error;
@@ -51,16 +51,6 @@ pub enum Error {
     CannotStopNode,
     #[error("Cannot send shutdown signal to the running node")]
     CannotSendStopSignal(#[source] io::Error),
-    #[error("unable to create the address")]
-    AddressCreate(#[source] io::Error),
-    #[error("Invalid address")]
-    InvalidAddress(#[source] std::string::FromUtf8Error),
-    #[error("No secret key, did you mean to create a secret key too?")]
-    NoSecretKey,
-    #[error("Unable to extract the public key")]
-    ReadPublicKey(#[from] io::Error),
-    #[error("Cannot generate key {0}")]
-    GenerateKey(String),
     #[error("REST is not running")]
     RestNotRunning,
 }
@@ -303,83 +293,6 @@ impl<'a> RunnerControl<'a> {
             Ok(())
         } else {
             Err(Error::CannotStopNode)
-        }
-    }
-
-    pub fn get_wallet_secret_key(&mut self, force: bool) -> Result<PathBuf, Error> {
-        let wallet_path = self.blockchain.get_wallet_secret();
-
-        if !wallet_path.is_file() || force {
-            self.gen_secret_key("Ed25519", &wallet_path)?;
-        }
-
-        Ok(wallet_path)
-    }
-
-    pub fn get_wallet_address(&mut self) -> Result<String, Error> {
-        let pk = self.get_public_key(self.blockchain.get_wallet_secret())?;
-
-        let address = self.make_address(pk.trim_end())?;
-
-        Ok(address.trim_end().to_owned())
-    }
-
-    fn make_address<PK: AsRef<str>>(&mut self, public_key: PK) -> Result<String, Error> {
-        let output = self
-            .jcli()
-            .args(&[
-                "address",
-                "account",
-                "--testing",
-                "--prefix=jorup_",
-                public_key.as_ref(),
-            ])
-            .output()
-            .map_err(Error::AddressCreate)?;
-        String::from_utf8(output.stdout).map_err(Error::InvalidAddress)
-    }
-
-    fn get_public_key<P>(&mut self, secret_key: P) -> Result<String, Error>
-    where
-        P: AsRef<Path>,
-    {
-        if !secret_key.as_ref().is_file() {
-            return Err(Error::NoSecretKey);
-        }
-
-        let output = self
-            .jcli()
-            .args(&[
-                "key",
-                "to-public",
-                "--input",
-                secret_key.as_ref().display().to_string().as_str(),
-            ])
-            .output()
-            .map_err(Error::ReadPublicKey)?;
-
-        String::from_utf8(output.stdout).map_err(Error::InvalidAddress)
-    }
-
-    fn gen_secret_key<P>(&mut self, key_type: &str, path: P) -> Result<(), Error>
-    where
-        P: AsRef<Path>,
-    {
-        let status = self
-            .jcli()
-            .args(&[
-                "key",
-                "generate",
-                "--type",
-                key_type,
-                path.as_ref().display().to_string().as_str(),
-            ])
-            .status()
-            .map_err(|_| Error::GenerateKey(key_type.to_owned()))?;
-        if status.success() {
-            Ok(())
-        } else {
-            Err(Error::GenerateKey(key_type.to_owned()))
         }
     }
 }
